+++ src/map/animation_functions/geohash.test.ts
@@ -1,0 +1,9 @@
+import updateMarkers from "./updateMarkers";
+
+test("updateMarkerst", () => {
+    const toChangeMarkers: any = [{"type":{},"key":"th","ref":null,"props":{"duration":300,"position":[23.03485420396439,47.59188850065745],"children":{"type":{},"key":null,"ref":null,"props":{"children":[{"type":"span","key":null,"ref":null,"props":{"children":"key: th"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"#aggregated: 134"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"lat: 23.03485420396439"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"lon: 47.59188850065745"},"_owner":null,"_store":{}}]},"_owner":null,"_store":{}}},"_owner":null,"_store":{}}];
+    const sourceMarkers: any =  [{"type":{},"key":"t","ref":null,"props":{"duration":300,"position":[19.268669954270678,-157.98994488168077],"children":{"type":{},"key":null,"ref":null,"props":{"children":[{"type":"span","key":null,"ref":null,"props":{"children":"key: t"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"#aggregated: 1187"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"lat: 19.268669954270678"},"_owner":null,"_store":{}},{"type":"br","key":null,"ref":null,"props":{},"_owner":null,"_store":{}},{"type":"span","key":null,"ref":null,"props":{"children":"lon: -157.98994488168077"},"_owner":null,"_store":{}}]},"_owner":null,"_store":{}}},"_owner":null,"_store":{}}];
+    const changedMarkers = updateMarkers(toChangeMarkers, sourceMarkers, 1);
+    expect(changedMarkers[0].key).toEqual('th')
+    expect(changedMarkers[0].props.position).toEqual([19.268669954270678, -157.98994488168077])
+})
